requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
anthropic>=0.52.0
chromadb>=1.0.0
pymupdf>=1.25.0
python-dotenv>=1.0.0
